+++ cmd/agnostic-ai/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"os"
 
@@ -12,6 +13,10 @@ var version = "0.58.0"
 func main() {
 	if err := cli.NewRootCmd(version).Execute(); err != nil {
 		fmt.Fprintln(os.Stderr, err)
+		var exitErr interface{ ExitCode() int }
+		if errors.As(err, &exitErr) {
+			os.Exit(exitErr.ExitCode())
+		}
 		os.Exit(1)
 	}
 }